Chat spike created and UI sketch added.
Tested the sending of messages using jQuery, code is small but allows me to
see that 'chatting' using JSON is easier and with little overhead (not really tested).

Updated the database file and added a small SQL file with some testing data, nothing fancy.
Added a PDF file with the scanned UI sketch.