Correcting lobby chat.
Chat messages in lobby are working again, properly ordered and with auto-scrolling.
Removed version name from config file, is is now placed in the site layout. An update to the game layout will be made when the game part is stable enough.
Corrected errors found while changing chat message code.